Style: Refactor Common Widgets
