+++ lib/view/controller/info_form/info_form_controller.dart
@@ -2,7 +2,6 @@ import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:iris_flutter/config/config.dart';
-import 'package:iris_flutter/view/page/Info_form/info_form_page.dart';
 
 class InfoFormController {
   RxList<XFile> images = <XFile>[].obs;
@@ -14,7 +13,7 @@ class InfoFormController {
   TextEditingController heightController = TextEditingController();
   TextEditingController weightController = TextEditingController();
   RxString location = Config.enterLocation.obs;
-  TextEditingController dressController = TextEditingController();
+  TextEditingController clothesController = TextEditingController();
   TextEditingController noteController = TextEditingController();
 
   // 생성 시간
@@ -27,7 +26,7 @@ class InfoFormController {
     print('print heightControllerText: ${heightController.text}');
     print('print weightControllerText: ${weightController.text}');
     print('print locationControllerText: ${location.value}');
-    print('print dressControllerText: ${dressController.text}');
+    print('print dressControllerText: ${clothesController.text}');
     print('print noteController: ${noteController.text}');
     //사진 저장 +
   }
